01_Detectar_Imagenes/requirements.txt
streamlit
openai
pillow
base64